project restructured, locationError alert added
