--- frontend/src/Components/Intro/Intro.jsx
@@ -1,12 +1,0 @@
-
-function Intro(){
-    return(
-        <>
-            <section className="centered">
-                <h1>Добро пожаловать в Weather App</h1>
-            </section>
-        </>
-    )
-}
-
-export default Intro;
